Rename `CandidateStrategy` to `CollisionStrategy` to clarify the abstraction
Now that identifying candidate pairs is done outside the strategies using
`groupByKey`, the name of `CandidateStrategy` is no longer accurate. What the
strategies actually do is convert the LSH signature from the hash tables into a
groupable key, in order to determine which points collide (i.e. hash to the same
bucket). `CollisionStrategy` is a more accurate name for the abstraction.
