+++ src/main/scala/com/github/karlhigley/spark/neighbors/collision/SimpleCollisionStrategy.scala
@@ -1,4 +1,4 @@
-package com.github.karlhigley.spark.neighbors.candidates
+package com.github.karlhigley.spark.neighbors.collision
 
 import scala.util.hashing.MurmurHash3
 
@@ -9,18 +9,18 @@ import org.apache.spark.storage.StorageLevel
 import com.github.karlhigley.spark.neighbors.lsh.{ BitSignature, HashTableEntry, IntSignature }
 
 /**
- * A very simple candidate identification strategy based on
- * an OR-construction between hash functions/tables
+ * A very simple collision strategy for candidate identification
+ * based on an OR-construction between hash functions/tables
  *
  * (See Mining Massive Datasets, Ch. 3)
  */
-private[neighbors] class SimpleCandidateStrategy extends CandidateStrategy with Serializable {
+private[neighbors] class SimpleCollisionStrategy extends CollisionStrategy with Serializable {
 
   /**
-   * Identify candidates by finding a signature match
-   * in any hash table.
+   * Convert hash tables into an RDD that is "collidable" using groupByKey.
+   * The new keys contain the hash table id, and a hashed version of the signature.
    */
-  def identify(hashTables: RDD[_ <: HashTableEntry[_]]): RDD[(Product, Point)] = {
+  def apply(hashTables: RDD[_ <: HashTableEntry[_]]): RDD[(Product, Point)] = {
     val entries = hashTables.map(entry => {
       // Arrays are mutable and can't be used in RDD keys
       // Use a hash value (i.e. an int) as a substitute
